fomo/requirements.txt
astroid==2.2.5   
atomicwrites==        1.3.0   
attrs         ==      19.1.0  
bcrypt          ==    3.1.7   
blinker           ==  1.4     
boto3    ==           1.9.174 
botocore   ==         1.12.174
cffi         ==       1.12.3  
Click          ==     7.0     
docutils         ==   0.14    
Flask              == 1.1.dev0
Flask-Admin==         1.5.3   
Flask-Bcrypt ==       0.7.1   
Flask-Login    ==     0.4.1   
Flask-Mail       ==   0.9.1   
Flask-SQLAlchemy   == 2.4.0   
Flask-WhooshAlchemy== 0.56    
Flask-WTF         ==  0.14.2  
importlib-metadata == 0.18    
isort             ==  4.3.20  
itsdangerous       == 1.1.0   
Jinja2            ==  2.10.1  
jmespath           == 0.9.4   
lazy-object-proxy  == 1.4.1   
MarkupSafe        ==  1.1.1   
mccabe            ==  0.6.1   
more-itertools     == 7.0.0   
packaging          == 19.0    
Pillow             == 6.0.0   
pip               ==  19.1.1  
pluggy            ==  0.12.0  
psycopg2          ==  2.8.3   
py                 == 1.8.0   
pycparser          == 2.19    
pycryptodome       == 3.8.2   
pylint            ==  2.3.1   
pylint-flask       == 0.6     
pylint-plugin-utils ==0.5     
pyparsing        ==   2.4.0   
pytest            ==  4.6.3   
python-dateutil   ==  2.8.0   
s3transfer        ==  0.2.1   
secret            ==  0.8     
setuptools         == 40.8.0  
six               ==  1.12.0  
SQLAlchemy        ==  1.3.5   
tabulate           == 0.8.3   
typed-ast          == 1.4.0   
urllib3            == 1.25.3  
wcwidth           ==  0.1.7   
Werkzeug          ==  0.15.4  
wheel              == 0.33.4  
Whoosh    ==          2.7.4   
wrapt       ==        1.11.2  
WTForms       ==      2.2.1   
WTForms-SQLAlchemy==  0.1     
zipp               == 0.5.1 